rename directory Neo4jDesktop->initialConnection
