+++ src/components/initialConnection/ShimConnectModal.js
