+++ src/components/initialConnection/Splash.js
