+++ src/components/initialConnection/Troubleshooting.js
